Renamed `internals.py` to `numpy.py`.
