+++ nengo/networks/tests/test_oscillator.py
@@ -4,7 +4,7 @@ import pytest
 
 import nengo
 from nengo.utils.functions import piecewise
-from nengo.utils.internals import rmse
+from nengo.utils.numpy import rmse
 from nengo.utils.testing import Plotter
