+++ nengo/utils/functions.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 import collections
 try:
     from collections import OrderedDict
