+++ nengo/utils/numpy.py
@@ -1,7 +1,7 @@
 """
-Helper functions for use inside the Nengo package.
+Extra functions to extend the capabilities of Numpy.
 """
-
+from __future__ import absolute_import
 import numpy as np
